move renderable methods to a concern
